Dog Responds To Touch
The dog now responds to touch commands and moves around the screen. In
this build, a lot of floats were replaced with ints to speed up
computation. Even so, it is very slow and may crash the emulator.

+++ SheepHerder/src/edu/lehman/android/domain/Animal.java
@@ -20,11 +20,11 @@ public abstract class Animal {
 		this.height = height;
 	}
 	
-	public void moveX(float moveX) {
+	public void moveX(int moveX) {
 		position.setX(position.getX() + moveX);
 	}
 
-	public void moveY(float moveY) {
+	public void moveY(int moveY) {
 		position.setY(position.getY() + moveY);
 	}
